Fixes #39232 - Update webhooks table to use TableIndexPage
