--- webpack/ForemanWebhooks/Routes/Webhooks/WebhooksIndexPage/Components/WebhooksTable/Components/Formatters/enabledCellFormatter.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import EnabledCell from '../EnabledCell';
-
-const enabledCellFormatter = () => value => <EnabledCell condition={value} />;
-
-export default enabledCellFormatter;
