--- webpack/ForemanWebhooks/Routes/Webhooks/WebhooksIndexPage/Components/WebhooksTable/Components/Formatters/nameToEditFormatter.js
@@ -1,19 +1,0 @@
-import React from 'react';
-import NameToEditCell from '../NameToEditCell';
-
-const nameToEditFormatter = (controllerPluralize, onClick) => (
-  value,
-  { rowData: { canEdit, id, name } }
-) => (
-  <NameToEditCell
-    active={canEdit}
-    id={id}
-    name={encodeURI(name)}
-    controller={controllerPluralize}
-    onClick={onClick}
-  >
-    {value}
-  </NameToEditCell>
-);
-
-export default nameToEditFormatter;
